Move rgbtools to rosettasciio
