+++ rosettasciio/rsciio/tests/utils/test_rgbtools.py
@@ -19,7 +19,7 @@
 
 import numpy as np
 
-import hyperspy.misc.rgb_tools as rt
+import rsciio.utils.rgb_tools as rt
 
 
 class TestRGBTools:
